updates naming, organization based on pr feedback
